case error flag to int type

+++ mysys/mf_iocache.cc
@@ -501,7 +501,7 @@ int _my_b_read(register IO_CACHE *info, unsigned char *Buffer, size_t Count)
   {
     if (Count)
     {
-      info->error= left_length;		/* We only got this many char */
+      info->error= static_cast<int> (left_length);	/* We only got this many char */
       return(1);
     }
     length=0;				/* Didn't read any chars */
@@ -1183,7 +1183,7 @@ read_append_buffer:
     info->append_read_pos += copy_len;
     Count -= copy_len;
     if (Count)
-      info->error = save_count - Count;
+      info->error = static_cast<int> (save_count - Count);
 
     /* Fill read buffer with data from write buffer */
     memcpy(info->buffer, info->append_read_pos,
